core+wasm: Step 4 static-check fixups (clippy, unused code, one lint)
Resolve every cargo clippy --workspace --all-targets --all-features
-D warnings finding surfaced by the type cutover: needless lifetimes
on resolve_key/resolve_text, SiteAddr/RuleSites visibility (SiteAddr
made pub since RuleSites — needed pub for bench access — holds it),
KeyIdx::get removed as genuinely dead, a few now-redundant u32 casts
left over from the Span narrowing, and several pre-existing
calibrate.rs style lints (collapsible match, sort_by, type
complexity, needless range loop, one dead struct field) surfaced
once the file compiled again.

cargo fmt --all --check still reports pre-existing drift unrelated
to this migration (confirmed identical in scale on the pre-migration
commit); left alone per the no-blanket-reformat rule. corpus.rs,
written from scratch in this migration, is fully rustfmt-clean.

+++ crates/wasm/src/lib.rs
@@ -415,8 +415,8 @@ fn project(target: &Corpus, findings: &[ssc_core::Finding]) -> Vec<Finding> {
                 sid: target.key(f.key_idx).to_string(),
                 code: f.code,
                 severity: f.severity,
-                start: u16.start as u32,
-                end: u16.end as u32,
+                start: u16.start,
+                end: u16.end,
                 score: f.score,
                 args: f.args.clone(),
             }
